Add support for static resource zip files
Previously the script couldn't handle static resource and static
resource zip files as those have different metadata file name formats
and locations. The folders also need to be transferred as a whole. As
part of the change the script can also handle changes to files within
subfolders in aura and lwc components.